requirements.txt
Django==4.1.6
django-ckeditor==6.5.1
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
gunicorn==20.1.0
ipython==8.8.0
pgcli==3.5.0
psycopg2==2.9.5
requests==2.28.2
sentry-sdk==1.16.0
whitenoise==6.3.0
